+++ src/main/java/kitchenpos/purgomalumclient/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.common;
+package kitchenpos.purgomalumclient;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
