+++ src/test/java/kitchenpos/application/FakePurgomalumClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.common.PurgomalumClient;
+import kitchenpos.purgomalumclient.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
