+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -7,8 +7,7 @@ import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.domain.*;
-import kitchenpos.common.PurgomalumClient;
+import kitchenpos.purgomalumclient.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
